+++ superset/assets/javascripts/explore/components/ChartContainer.jsx
