+++ superset/assets/javascripts/explore/components/Control.jsx
