+++ superset/assets/javascripts/explore/components/ControlHeader.jsx
