+++ superset/assets/javascripts/explore/components/ControlPanelSection.jsx
