+++ superset/assets/javascripts/explore/components/ControlPanelsContainer.jsx
