+++ superset/assets/javascripts/explore/components/DisplayQueryButton.jsx
