+++ superset/assets/javascripts/explore/components/controls/CheckboxControl.jsx
