+++ superset/assets/javascripts/explore/components/controls/Filter.jsx
