+++ superset/assets/javascripts/explore/components/controls/FilterControl.jsx
