+++ superset/assets/javascripts/explore/components/controls/HiddenControl.jsx
