+++ superset/assets/javascripts/explore/components/controls/SelectControl.jsx
