+++ superset/assets/javascripts/explore/components/controls/TextAreaControl.jsx
