+++ superset/assets/javascripts/explore/components/controls/TextControl.jsx
