+++ superset/assets/javascripts/explore/exploreUtils.js
