+++ superset/assets/javascripts/explore/reducers/exploreReducer.js
