+++ superset/assets/javascripts/explore/stores/controls.jsx
