+++ superset/assets/javascripts/explore/stores/visTypes.js
