+++ superset/assets/javascripts/explore/validators.js
