+++ superset/assets/spec/javascripts/explorev2/components/FilterControl_spec.jsx
@@ -5,8 +5,8 @@ import sinon from 'sinon';
 import { expect } from 'chai';
 import { describe, it, beforeEach } from 'mocha';
 import { shallow } from 'enzyme';
-import FilterControl from '../../../../javascripts/explorev2/components/controls/FilterControl';
-import Filter from '../../../../javascripts/explorev2/components/controls/Filter';
+import FilterControl from '../../../../javascripts/explore/components/controls/FilterControl';
+import Filter from '../../../../javascripts/explore/components/controls/Filter';
 
 const defaultProps = {
   choices: ['country_name'],
