+++ superset/assets/spec/javascripts/explorev2/components/URLShortLinkButton_spec.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { expect } from 'chai';
 import { describe, it } from 'mocha';
 
-import URLShortLinkButton from '../../../../javascripts/explorev2/components/URLShortLinkButton';
+import URLShortLinkButton from '../../../../javascripts/explore/components/URLShortLinkButton';
 
 describe('URLShortLinkButton', () => {
   const defaultProps = {
