+++ superset/assets/spec/javascripts/sqllab/VisualizeModal_spec.jsx
@@ -14,7 +14,7 @@ import { queries } from './fixtures';
 import { sqlLabReducer } from '../../../javascripts/SqlLab/reducers';
 import * as actions from '../../../javascripts/SqlLab/actions';
 import VisualizeModal from '../../../javascripts/SqlLab/components/VisualizeModal';
-import * as exploreUtils from '../../../javascripts/explorev2/exploreUtils';
+import * as exploreUtils from '../../../javascripts/explore/exploreUtils';
 
 global.notify = {
   info: () => {},
