+++ superset/templates/superset/explore.html
@@ -17,7 +17,7 @@
 
 {% block tail_js %}
   {{ super() }}
-  {% with filename="explorev2" %}
+  {% with filename="explore" %}
     {% include "superset/partials/_script_tag.html" %}
   {% endwith %}
 {% endblock %}
